a option of "tracer" added. follow the logical parameter itracer set in SPECTRAL.f90. by default it is turned off.
  : I M P O R T A N T  N O T E :
Ideally when the tracer is turned off, the solution must be exact to the solution produced
by the code that did not have the tracer.
However, they are different (to the order of 10^-6).

For some reason, if the calls to time advance routines for the tracer are commented (when
the tracer is off) ie. the call tradvB and tradvT are commented, the solutions are exact.




git-svn-id: https://svn.riouxsvn.com/sqgmodel/trunk@3 bde0f0f0-3b85-4903-a1d0-8e0114a44ce5