M    fortran/sqg_mod.f90 writes basic state to output file or as a seperate file
M    fortran/spectral_mod.f90
added a flag ibase, to write basic state to the output file or to disk as a seperate file

M    fortran/Makefile
flags are optimized for compiling

M    fortran/pvinv_mod.f90
edited the warnings that were occurring in matinv.
also, if the base state is in the output file, it can be removed with the help of ibase flag.




git-svn-id: https://svn.riouxsvn.com/sqgmodel/trunk@61 bde0f0f0-3b85-4903-a1d0-8e0114a44ce5